requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
chromedriver-binary==74.0.3729.6.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
exceptiongroup==1.1.1
facebook-sdk==3.1.0
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
mccabe==0.7.0
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pandas==2.0.1
pathspec==0.11.1
pathy==0.10.1
platformdirs==3.2.0
preshed==3.0.8
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
Pympler==1.0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.9.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
textblob==0.17.1
thinc==8.1.9
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vaderSentiment==3.3.2
wasabi==1.1.1
wsproto==1.2.0
